refactor(directory): step1 - 디렉토리 구조 변경(repository 추가)
