+++ src/main/java/qna/repository/UserRepository.java
@@ -1,8 +1,9 @@
-package qna.domain;
+package qna.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
+import qna.domain.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUserId(String userId);
